Add more headers for ending to send with https requests
When trying to update data in the admin panel, I was getting a "Request
origin does not match request base_url" error. This is because when
nginx is acting as a reverse proxy (which it is my case) rack apps
expect some extra headers that I wasn't sending.

I found the solution here: https://stackoverflow.com/a/51111144/638966