Only run asdf install if needed
When running asdf install I was running into a weird thing where it
seemed like it was installing even if it didn't need to (based on the
fact that the step was taking a long time). To prevent this, I added a
status check. A new step now runs `asdf current`. If it finds any lines
that contain "Not installed" it will run `asdf install`.

Also `asdf install` is idempotent (when running correctly) so this
shouldn't be required. However, ansible won't know if anything changed.
To give it the correct signal, I can also use the asdf status step to
know if it actually tried to install something, which is good practice.