Underline all links
For accessibility. [Understanding WCAG
2.0](https://www.w3.org/TR/UNDERSTANDING-WCAG20/Overview.html),
[Success Criterion
1.4.1](https://www.w3.org/TR/UNDERSTANDING-WCAG20/visual-audio-contrast-
without-color.html) notes that color should not be the only way to
distinguish an element on the page. For example what if the website is
viewed on a black and white eInk screen, or printed out, or by someone
with color blindness, or using the greyscale mode on their device.

My links were mostly not underlined and only used color to appear. I
added underlines as they are widely understood to communicate "this is
a link, buddy".