Create drag and drop with stimulus
This is a big one.

I wanted to remove jQuery from the site. No offense to jQuery, it defined a large part of the web for many years, but I think we've moved beyond it now. I also wanted to get some closer hands on experience with Stimulus.

This details how I created drag and drop functionality to rearrange elements.

Drag and drop is relatively broad set of functionality. You can drag in files from outside the browser, drag elements from group to group (changing or categorization), put items inside one another (changing their parent), and more. I'm doing something else: changing their vertical position in a group. In this case, changing the order of projects or songs. I also wanted to ensure the drag was cancelable and show where the element would end up if you drop it.

The way stimulus works is actions on an element convert javascript events into controller methods. The drag and drop events to track are:

* `dragstart`
* `dragenter`
* `dragover`
* `drop`
* `dragend`

There's also `drag` `dragleave` but I didn't end up needing those.

The DragItemsController keeps track of where the dragged item is being dropped so that it can show the preview element in the correct spot. It appears to be convention to set some drag data when the drag starts, so I store the data grabbed from the `data-drag-items-id` attribute on each `li`. This is also used to exclude child elements as drop targets. Because these list elements have child elements, they could be valid drop targets, but I don't want that. Every time DragItemsController evaluates a drop target, it checks to see if it has a `data-drag-items-id` attribute. If not, it rejects that target.

Since elements can be inserted before or after any given element in the list, the DragItemsController also keeps track of that before or after state. It calculates the mouse's Y position and compares it to the drop target element's middle. That is, take the result of `getBoundingClientRect()`, get the rect's `top`, and add the `height` of the element divided by 2. If the mouse is below the middle, insert the element after the dropped element. Otherwise, insert above.

Once the element is dropped (and is successful) it inserts the element in the correct spot and dispatches the `success` event. This can be used to sync the element order on the server, or whatever else.

The preview item's class can be set for styling in the `data-drag-items-preview-class` attribute on the container level.

## References

https://onrails.blog/2018/03/09/stimulus-js-tutorial-how-do-i-drag-and-drop-items-in-a-list/
https://developer.mozilla.org/en-US/docs/Web/API/HTML_Drag_and_Drop_API
https://developer.mozilla.org/en-US/docs/Web/API/HTML_Drag_and_Drop_API/Drag_operations