Sync the sorted items
Introducing a SortItems controller that takes a container of items with
a `data-sort-item-id` and sends a POST request to a url with the those
ids in order. It also grabs the csrf token from the `<head>` and sets
it as a POST header.

## Controller

`sort-items` is expected to exist on the a container with the items to
be sorted inside, e.g. the `ul` containing sortable `li`s

## Values

* sort-key: The key to hold all the IDs in the POST payload. E.g.
`projects` will send a payload that like: `{ "project": ["1", "2", "3"]
}`
* url: Where to send the POST request

## Other info

* `data-sort-item-id` attribute: on each sortable item; put the ID of
the item here so the server can put the items with those IDs in that
order.